refactor BasejailStorage
